fix typo: use int instead of double for duplicate point detection
No change in output expected, but maybe a minimal performance impact

git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@3819 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/util/Java2DConverter.java
@@ -89,7 +89,7 @@ public class Java2DConverter {
 			// because we use closePath() to signal that
 			--n;
 		}
-		double lastLat = Integer.MAX_VALUE,lastLon = Integer.MAX_VALUE;
+		int lastLat = Integer.MAX_VALUE,lastLon = Integer.MAX_VALUE;
 		for (int i = 0; i < n; i++){
 			Coord co = polygonPoints.get(i);
 			int latHp = co.getHighPrecLat();
